Completed tasks and tried to bug fix loading bar.
