+++ public/index.html
@@ -4,12 +4,14 @@
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
     <meta
-      name="Front End Developer Intern Challenge for Summer 2022 | Linda Ngoc Nguyen"
-      content="An awesome example of using NASA's API to display the best space-related astronomy images."
+      name="description"
+      content="Front End Developer Intern Challenge for Summer 2022 | Linda Ngoc Nguyen | An awesome example of using NASA's API to display the best space-related astronomy images."
     />
+    <meta name="author" content="Linda Ngoc Nguyen">
+    <meta name="keywords" content="React, JavaScript, HTML, CSS, Tailwind CSS"
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <title>Front End Developer Intern Challenge for Summer 2022 - Spaceroid</title>
   </head>
   <body>
     <div id="root"></div>
