+++ src/index.js
@@ -1,15 +1,31 @@
+/* eslint-disable no-unused-vars */
 /* eslint-disable react/jsx-filename-extension */
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter as Router } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Navigation from "./components/navigation";
+import Hero from "./pages/hero";
+import Main from "./pages/main";
 import "./styles/global.css";
 
 ReactDOM.render(
+  // <Router>
+  //   <Navigation />
+  //   <Header />
+  //   <Main />
+  //   <Switch>
+  //     <Route exact path="/" element={<Header />} />
+  //     <Route path="/main/" element={<Main />} />
+  //   </Switch>
+  // </Router>,
   <div>
-    <Router>
+    <BrowserRouter>
       <Navigation />
-    </Router>
+      <Routes>
+        <Route exact path="/" element={<Hero />} />
+        <Route path="/main/" element={<Main />} />
+      </Routes>
+    </BrowserRouter>
   </div>,
   document.getElementById("root")
 );
