+++ src/pages/hero.js
@@ -3,33 +3,30 @@ import { Link } from "react-router-dom";
 import Astronaut from "../images/astronaut.svg";
 import Logo from "../images/logo.svg";
 
-export default function hero() {
+export default function Hero() {
   return (
-    <main
-      id="hero"
-      className="fixed h-screen flex flex-col z-background text-center"
-    >
+    <main className="flex flex-col justify-center text-center z-background">
       <img
+        id="hero"
         src={Astronaut}
-        className="h-5/6 z-background"
+        className="h-[calc(100vh-30px)] rotate-180"
         alt="An astronaut standing on a reddish purple planet that is upside-down in space."
       />
-      <img
-        src={Logo}
-        className="h-logoHeroHeight sm:mt-marginHeroLogo mobile:mt-marginHeroLogoMobile z-1"
-        alt="Spaceroid Logo"
-      />
-      <h1 className="m-3">
-        The finest collection of space-related images on the internet.
-      </h1>
-      <Link to="/images">
-        <li
-          key="images"
-          className="button inline p-3 px-5 bg-shootingStar rounded-lg"
-        >
-          Start Browsing
-        </li>
-      </Link>
+      <div className="tagline w-full fixed flex flex-col z-10">
+        <img src={Logo} className="h-logoHeroHeight" alt="Spaceroid Logo" />
+        <h1 className="m-3">
+          The finest collection of space-related images on the internet.
+        </h1>
+        <Link to="/main">
+          <button
+            key="images"
+            type="button"
+            className="button inline p-3 px-5 bg-shootingStar rounded-lg"
+          >
+            Start Browsing
+          </button>
+        </Link>
+      </div>
     </main>
   );
 }
