+++ src/pages/main.js
@@ -1,0 +1,11 @@
+import React from "react";
+import Section from "../components/section";
+import "../styles/global.css";
+
+export default function Main() {
+  return (
+    <main>
+      <Section />
+    </main>
+  );
+}
