+++ tailwind.config.js
@@ -15,15 +15,15 @@ module.exports = {
         logoNavWidth: "140px",
         logoHeroWidth: "240px",
         likesWidth: "40px",
-        spinnerWidth: "150px",
+        spinnerWidth: "100px",
         imageWidth: "460px",
         imageWidthMobile: "300px",
       },
       height: {
-        logoNavHeight: "65px",
+        logoNavHeight: "30px",
         logoHeroHeight: "60px",
         likesHeight: "30px",
-        spinnerHeight: "150px",
+        spinnerHeight: "100px",
         imageHeight: "460px",
         imageHeightMobile: "300px",
       },
@@ -31,12 +31,14 @@ module.exports = {
         background: "-10",
       },
       spacing: {
-        marginHeroLogo: "-11%",
-        marginHeroLogoMobile: "-20%",
-        imagesTop: "120px",
+        imagesTop: "50px",
       },
       colors: {
-        shootingStar: "#FC636B",
+        shootingStar: "#E3414A",
+        fallingStar: "#9E1A21",
+      },
+      animation: {
+        slow: "spin 10s linear infinite",
       },
     },
   },
